+++ src/schemas/index_filter_values_schema.ts
@@ -1,0 +1,11 @@
+export const indexFilterValuesSchema = {
+    id: '/indexFilterValues',
+    additionalProperties: {
+        oneOf: [
+            {$ref: '/numberSchema'},
+            {$ref: '/addressSchema'},
+            {$ref: '/orderHashSchema'},
+        ],
+    },
+    type: 'object',
+};