+++ src/schemas/subscription_opts_schema.ts
@@ -1,0 +1,20 @@
+export const blockParamSchema = {
+    id: '/blockParam',
+    oneOf: [
+        {
+            type: 'number',
+        },
+        {
+            enum: ['latest', 'earliest', 'pending'],
+        },
+    ],
+};
+
+export const subscriptionOptsSchema = {
+    id: '/subscriptionOpts',
+    properties: {
+        fromBlock: {$ref: '/blockParam'},
+        toBlock: {$ref: '/blockParam'},
+    },
+    type: 'object',
+};