+++ src/utils/schema_validator.ts
@@ -1,8 +1,11 @@
 import {Validator, ValidatorResult,  Schema} from 'jsonschema';
 import {ecSignatureSchema, ecSignatureParameterSchema} from '../schemas/ec_signature_schema';
+import {orderHashSchema} from '../schemas/order_hash_schema';
 import {orderSchema, signedOrderSchema} from '../schemas/order_schemas';
 import {addressSchema, numberSchema} from '../schemas/basic_type_schemas';
 import {tokenSchema} from '../schemas/token_schema';
+import {subscriptionOptsSchema, blockParamSchema} from '../schemas/subscription_opts_schema';
+import {indexFilterValuesSchema} from '../schemas/index_filter_values_schema';
 import {orderFillOrKillRequestsSchema} from '../schemas/order_fill_or_kill_requests_schema';
 
 export class SchemaValidator {
@@ -13,8 +16,12 @@ export class SchemaValidator {
         this.validator.addSchema(orderSchema, orderSchema.id);
         this.validator.addSchema(numberSchema, numberSchema.id);
         this.validator.addSchema(addressSchema, addressSchema.id);
+        this.validator.addSchema(orderHashSchema, orderHashSchema.id);
+        this.validator.addSchema(blockParamSchema, blockParamSchema.id);
         this.validator.addSchema(ecSignatureSchema, ecSignatureSchema.id);
         this.validator.addSchema(signedOrderSchema, signedOrderSchema.id);
+        this.validator.addSchema(subscriptionOptsSchema, subscriptionOptsSchema.id);
+        this.validator.addSchema(indexFilterValuesSchema, indexFilterValuesSchema.id);
         this.validator.addSchema(ecSignatureParameterSchema, ecSignatureParameterSchema.id);
         this.validator.addSchema(orderFillOrKillRequestsSchema, orderFillOrKillRequestsSchema.id);
     }